requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==4.0.1
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.66.0
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.2.2
MarkupSafe==2.1.5
msgpack==1.0.8
nltk==3.8.1
numpy==2.0.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
zeep==4.2.1
